requirements.txt
streamlit>=1.28.0
databricks-sdk>=0.18.0
pandas>=1.5.0
pillow>=9.0.0
python-dotenv>=1.0.0